dcad-scraper-with-api/dcad-scraper/scraper/requirements.txt
playwright==1.47.0
beautifulsoup4==4.12.3
lxml==5.2.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
pydantic==2.8.2
python-dotenv==1.0.1
fastapi==0.111.1
uvicorn[standard]==0.30.6
